refactor(site): move task mutations to api/queries/tasks
Move pauseTask/resumeTask mutation configs from useTaskActions.ts hook
wrappers into api/queries/tasks.ts, following the codebase convention of
defining mutation configs in api/queries/ and using useMutation inline.
