+++ site/src/api/queries/tasks.ts
@@ -1,0 +1,36 @@
+import { API } from "api/api";
+import type { Task } from "api/typesGenerated";
+import type { QueryClient } from "react-query";
+
+export const pauseTask = (task: Task, queryClient: QueryClient) => {
+	return {
+		mutationFn: async () => {
+			if (!task.workspace_id) {
+				throw new Error("Task has no workspace");
+			}
+			return API.stopWorkspace(task.workspace_id);
+		},
+		onSuccess: async () => {
+			await queryClient.invalidateQueries({ queryKey: ["tasks"] });
+		},
+	};
+};
+
+export const resumeTask = (task: Task, queryClient: QueryClient) => {
+	return {
+		mutationFn: async () => {
+			if (!task.workspace_id) {
+				throw new Error("Task has no workspace");
+			}
+			return API.startWorkspace(
+				task.workspace_id,
+				task.template_version_id,
+				undefined,
+				undefined,
+			);
+		},
+		onSuccess: async () => {
+			await queryClient.invalidateQueries({ queryKey: ["tasks"] });
+		},
+	};
+};
