+++ site/src/modules/tasks/taskActions.ts
@@ -3,7 +3,7 @@ import type { TaskStatus } from "api/typesGenerated";
 /**
  * Task statuses that allow pausing.
  */
-const pauseStatuses: TaskStatus[] = [
+const PAUSABLE_STATUSES: TaskStatus[] = [
 	"active",
 	"initializing",
 	"pending",
@@ -14,30 +14,30 @@ const pauseStatuses: TaskStatus[] = [
 /**
  * Task statuses where the pause button should be disabled (in transition).
  */
-const pauseDisabledStatuses: TaskStatus[] = ["pending", "initializing"];
+const PAUSE_DISABLED_STATUSES: TaskStatus[] = ["pending", "initializing"];
 
 /**
  * Task statuses that allow resuming.
  */
-const resumeStatuses: TaskStatus[] = ["paused", "error", "unknown"];
+const RESUMABLE_STATUSES: TaskStatus[] = ["paused", "error", "unknown"];
 
 /**
  * Checks if a task can be paused based on its status.
  */
 export function canPauseTask(status: TaskStatus): boolean {
-	return pauseStatuses.includes(status);
+	return PAUSABLE_STATUSES.includes(status);
 }
 
 /**
  * Checks if the pause action should be disabled for a task status.
  */
 export function isPauseDisabled(status: TaskStatus): boolean {
-	return pauseDisabledStatuses.includes(status);
+	return PAUSE_DISABLED_STATUSES.includes(status);
 }
 
 /**
  * Checks if a task can be resumed based on its status.
  */
 export function canResumeTask(status: TaskStatus): boolean {
-	return resumeStatuses.includes(status);
+	return RESUMABLE_STATUSES.includes(status);
 }
